Move textures to new file
