+++ src/client/client.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "client_engine.h"
-#include "gl/gl_object.h"
 #include "gl/shader.h"
+#include "gl/textures.h"
 #include "gl/vertex_array.h"
 #include "maths.h"
 #include "world/chunk_mesh.h"
